Add files via upload
Code is given in  a Jupyter NoteBook. And the images can be downloiaded from Kaggle dataset. 
The implementation of the Wavelet CNN (in the manuscript) is done with Functiona API to seek Spatial domain images to the CNN i/p and Wavelet sub- bands (directly) to the Dense layer. The design uniquness to accamodate MRA in a CNN. Custom activation function is generated and called in the Model built. All the modules in the manuscript is given below except 1. Weighted and Combined loss 2. Heat map visualization of Image and the Wavelet sub- bands, 3. Visualization & Evaluation of intermediate activation maps